frontend/requirements.txt
streamlit
requests
numpy
plotly.express